Move sprite component to be game agnostic
EU5 trade goods uses the sprite component, so no need to house it in EU4
any longer
