+++ src/app/app/features/eu4/components/avatars/BuildingAvatar.tsx
@@ -1,7 +1,7 @@
 import type { GfxObj } from "@/features/eu4/types/models";
 import { Tooltip } from "@/components/Tooltip";
-import { Sprite, spriteDimension } from "../Sprite";
-import type { SpriteDimension } from "../Sprite";
+import { Sprite, spriteDimension } from "@/components/Sprite";
+import type { SpriteDimension } from "@/components/Sprite";
 import globalSpriteData from "@/images/eu4/buildings/global.json";
 import westernSpriteData from "@/images/eu4/buildings/westerngfx.json";
 import globalSpriteSrc from "@/images/eu4/buildings/global.webp";
