+++ src/app/app/features/eu4/components/avatars/TCInvestmentAvatar.tsx
@@ -1,7 +1,7 @@
 import type { LocalizedObj } from "@/features/eu4/types/models";
 import { Tooltip } from "@/components/Tooltip";
-import { Sprite, spriteDimension } from "../Sprite";
-import type { SpriteDimension } from "../Sprite";
+import { Sprite, spriteDimension } from "@/components/Sprite";
+import type { SpriteDimension } from "@/components/Sprite";
 import investmentData from "@/images/eu4/tc-investments/investments.json";
 import investmentImage from "@/images/eu4/tc-investments/investments.webp";
